feat: add adding courses
