+++ frontend/src/pages/Home.module.css
@@ -200,3 +200,22 @@
   color: #237723;
   cursor: pointer;
 }
+
+.courseFormSection {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  margin-top: 20px;
+}
+
+.courseFormSection p {
+  text-align: left;
+}
+
+.courseInput {
+  border: 1px solid #0c1e66;
+  box-sizing: border-box;
+  padding: 10px;
+  width: 100%;
+  background-color: #fcfcfc;
+}
